Condition for editing conf file

+++ images/9.0.4-php7.3/docker-run.sh
@@ -14,21 +14,23 @@ function initDolibarr()
   sendmail_path = /usr/sbin/sendmail -t -i
 EOF
 
-  cat > /var/www/html/conf/conf.php << EOF
-  <?php
-  \$dolibarr_main_url_root='${DOLI_URL_ROOT}';
-  \$dolibarr_main_document_root='/var/www/html';
-  \$dolibarr_main_url_root_alt='/custom';
-  \$dolibarr_main_document_root_alt='/var/www/html/custom';
-  \$dolibarr_main_data_root='/var/www/documents';
-  \$dolibarr_main_db_host='${DOLI_DB_HOST}';
-  \$dolibarr_main_db_port='3306';
-  \$dolibarr_main_db_name='${DOLI_DB_NAME}';
-  \$dolibarr_main_db_prefix='llx_';
-  \$dolibarr_main_db_user='${DOLI_DB_USER}';
-  \$dolibarr_main_db_pass='${DOLI_DB_PASSWORD}';
-  \$dolibarr_main_db_type='mysqli';
+  if [ ! -f /var/www/html/conf/conf.php ]; then
+    cat > /var/www/html/conf/conf.php << EOF
+    <?php
+    \$dolibarr_main_url_root='${DOLI_URL_ROOT}';
+    \$dolibarr_main_document_root='/var/www/html';
+    \$dolibarr_main_url_root_alt='/custom';
+    \$dolibarr_main_document_root_alt='/var/www/html/custom';
+    \$dolibarr_main_data_root='/var/www/documents';
+    \$dolibarr_main_db_host='${DOLI_DB_HOST}';
+    \$dolibarr_main_db_port='3306';
+    \$dolibarr_main_db_name='${DOLI_DB_NAME}';
+    \$dolibarr_main_db_prefix='llx_';
+    \$dolibarr_main_db_user='${DOLI_DB_USER}';
+    \$dolibarr_main_db_pass='${DOLI_DB_PASSWORD}';
+    \$dolibarr_main_db_type='mysqli';
 EOF
+  fi
 
   chown -R www-data:www-data /var/www
   chmod 400 /var/www/html/conf/conf.php
